Ajustes

+++ src/main/java/br/com/fiap/project/repository/NotificacaoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NotificacaoRepository extends JpaRepository<Notificacao, Long> {
+public interface NotificacaoRepository extends JpaRepository<Notificacao, Integer> {
     // Métodos de consulta personalizados, se necessário
 }
